webappnew/requirements.txt
Flask==2.3.3
python-dotenv
loguru
yookassa
requests
aiosqlite
py3xui
aiohttp